requirements.txt
python-telegram-bot
dotenv
mistralai
python-telegram-bot[job-queue]